+++ flake.nix
@@ -1,5 +1,5 @@
 {
-  description = "A Nix-flake-based Java development environment";
+  description = "Fray Flake Configuration";
 
   inputs = {
     nixpkgs.url = "https://flakehub.com/f/NixOS/nixpkgs/0.1.*.tar.gz";
@@ -12,8 +12,6 @@
     ,
     }:
     let
-      javaVersion = 23; # Change this value to update the whole stack
-
       supportedSystems = [ "x86_64-linux" "aarch64-darwin" ];
       forEachSupportedSystem = f:
         nixpkgs.lib.genAttrs supportedSystems (system:
@@ -25,31 +23,59 @@
           });
     in
     {
-      overlays.default = final: prev:
-        let
-          jdk = prev."jdk${toString javaVersion}";
-        in
-        {
-          inherit jdk;
-          gradle = prev.gradle.override { java = jdk; };
-        };
+      overlays.default = final: prev: {
+        jdk = prev.jdk23;
+        java = prev.jdk23;
+      };
 
       packages = forEachSupportedSystem (
         { pkgs }:
         let
-          project = pkgs.gradle2nix.buildGradlePackage {
-            pname = "fray";
-            version = "0.4.0";
+          project =
+            (gradle2nix.builders.${pkgs.system}.buildGradlePackage {
+              pname = "fray";
+              version = "0.4.5-SNAPSHOT";
 
-            src = ./.;
-
-            nativeBuildInputs = with pkgs; [
-              jdk
-              gradle
-            ];
-
-            deps = ./deps.json;
-          };
+              src = ./.;
+              lockFile = ./gradle.lock;
+              gradleBuildFlags = [
+                "-Porg.gradle.java.installations.paths=${pkgs.jdk11.home},${pkgs.jdk23.home}"
+                "build"
+                "-x"
+                "test"
+              ];
+              buildInputs = with pkgs; [
+                jdk23
+                jdk11
+                gcc
+                cmake
+              ];
+              preBuild = ''
+                export CC="${pkgs.gcc}/bin/gcc"
+                export CXX="${pkgs.gcc}/bin/g++"
+                export JDK11="${pkgs.jdk11.home}"
+                export JRE="${pkgs.jdk23.home}"
+                export JAVA_HOME="${pkgs.jdk23.home}"
+                ${pkgs.lib.optionalString pkgs.stdenv.isLinux ''
+                  export JETBRAINS_JDK_HOME="${pkgs.jetbrains.jdk.home}"
+                ''}
+                sed -i '/include("plugins/d' settings.gradle.kts
+              '';
+            }).overrideAttrs (_: prev: {
+              gradleFlags = pkgs.lib.lists.remove "--console=plain" prev.gradleFlags;
+              installPhase = ''
+                runHook preInstall
+                mkdir -p $out/libs
+                mkdir -p $out/bin
+                cp core/build/libs/*.jar $out/libs
+                cp instrumentation/agent/build/libs/*.jar $out/libs
+                cp -r instrumentation/jdk/build/java-inst $out/
+                cp -r jvmti/build/native-libs $out/
+                gradle --no-daemon -Pfray.installDir=$out/ genRunner
+                cp bin/fray $out/bin/
+                runHook postInstall
+              '';
+            });
         in
         {
           default = project;
@@ -63,9 +89,8 @@
               gcc
               cmake
               gradle
-              jdk
+              jdk23
               jdk11
-              jdk21
             ]
             ++ lib.optionals (pkgs.stdenv.isLinux) [
               jetbrains.jdk
@@ -74,9 +99,8 @@
             export CC="${pkgs.gcc}/bin/gcc"
             export CXX="${pkgs.gcc}/bin/g++"
             export JDK11="${pkgs.jdk11.home}"
-            export JDK21="${pkgs.jdk21.home}"
-            export JRE="${pkgs.jdk.home}"
-            export JAVA_HOME="${pkgs.jdk.home}"
+            export JRE="${pkgs.jdk23.home}"
+            export JAVA_HOME="${pkgs.jdk23.home}"
             ${pkgs.lib.optionalString pkgs.stdenv.isLinux ''
               export JETBRAINS_JDK_HOME="${pkgs.jetbrains.jdk.home}"
             ''}